@@ -1,9 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
         <Description>Entity Framework Core Second Level Caching Library.</Description>
-        <VersionPrefix>4.8.7</VersionPrefix>
+        <VersionPrefix>4.8.8</VersionPrefix>
         <Authors>Vahid Nasiri</Authors>
-        <TargetFrameworks>net8.0;net7.0;net6.0;net5.0;netstandard2.1;netstandard2.0;net462;netcoreapp3.1;</TargetFrameworks>
+        <TargetFrameworks>net9.0;net8.0;net7.0;net6.0;net5.0;netstandard2.1;netstandard2.0;net462;netcoreapp3.1;</TargetFrameworks>
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
         <AssemblyName>EFCoreSecondLevelCacheInterceptor</AssemblyName>
         <PackageId>EFCoreSecondLevelCacheInterceptor</PackageId>
@@ -48,11 +48,11 @@
         </WarningsNotAsErrors>
     </PropertyGroup>
     <ItemGroup>
-        <PackageReference Include="Meziantou.Analyzer" Version="2.0.176">
+        <PackageReference Include="Meziantou.Analyzer" Version="2.0.179">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.11.20">
+        <PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.12.19">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
@@ -157,6 +157,17 @@
     </ItemGroup>
     <PropertyGroup Condition="('$(TargetFramework)' == 'net8.0')">
         <DefineConstants>NET8_0</DefineConstants>
+    </PropertyGroup>
+
+    <ItemGroup Condition="('$(TargetFramework)' == 'net9.0')">
+        <PackageReference Include="System.IO.Hashing" Version="[9.0.0,)"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="[9.0.0,)"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="[9.0.0,)"/>
+        <PackageReference Include="CacheManager.Core" Version="[1.2,)"/>
+        <PackageReference Include="EasyCaching.Core" Version="[1.9.2,)"/>
+    </ItemGroup>
+    <PropertyGroup Condition="('$(TargetFramework)' == 'net9.0')">
+        <DefineConstants>NET9_0</DefineConstants>
     </PropertyGroup>
 
     <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
