@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk.Worker">
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <UserSecretsId>dotnet-EFCoreSecondLevelCacheInterceptor.WorkerServiceWithLamar-7BDED8C3-74B4-4CA0-82AA-E5452A3AF4A7</UserSecretsId>
   </PropertyGroup>
   <ItemGroup>
@@ -8,11 +8,11 @@
     <ProjectReference Include="..\EFCoreSecondLevelCacheInterceptor.Tests.DataLayer\EFCoreSecondLevelCacheInterceptor.Tests.DataLayer.csproj" />
   </ItemGroup>  
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.0" />
     <PackageReference Include="Lamar.Microsoft.DependencyInjection" Version="13.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.1" />
-    <PackageReference Include="Microsoft.Extensions.Options" Version="8.0.2" /> 
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.2" /> 
-    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="8.0.1" /> 
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Options" Version="9.0.0" /> 
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.0" /> 
+    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="9.0.0" /> 
   </ItemGroup>
 </Project>
