@@ -6,17 +6,15 @@ namespace EFCoreSecondLevelCacheInterceptor;
 /// <summary>
 ///     CacheAllQueries Options
 /// </summary>
-public class CacheSpecificQueriesOptions : CacheAllQueriesOptions
+/// <remarks>
+///     CacheAllQueries Options
+/// </remarks>
+public class CacheSpecificQueriesOptions(IList<Type>? entityTypes) : CacheAllQueriesOptions
 {
-    /// <summary>
-    ///     CacheAllQueries Options
-    /// </summary>
-    public CacheSpecificQueriesOptions(IList<Type>? entityTypes) => EntityTypes = entityTypes;
-
     /// <summary>
     ///     Given entity types to cache
     /// </summary>
-    public IList<Type>? EntityTypes { get; }
+    public IList<Type>? EntityTypes { get; } = entityTypes;
 
     /// <summary>
     ///     How should we determine which tables should be cached?
