@@ -1,5 +1,5 @@
 using System;
-#if NET5_0 || NET6_0 || NET7_0 || NET8_0
+#if NET9_0 || NET5_0 || NET6_0 || NET7_0 || NET8_0
 using System.Text.Json;
 #endif
 
@@ -12,7 +12,7 @@ public class EFCoreSecondLevelCacheOptions
 {
     internal EFCoreSecondLevelCacheSettings Settings { get; } = new();
 
-#if NET5_0 || NET6_0 || NET7_0 || NET8_0
+#if NET9_0 || NET5_0 || NET6_0 || NET7_0 || NET8_0
     /// <summary>
     ///     Provides options to control the serialization behavior.
     ///     EFCacheKeyProvider uses these options to serialize the parameter values.
