@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-#if NET5_0 || NET6_0 || NET7_0 || NET8_0
+#if NET9_0 || NET5_0 || NET6_0 || NET7_0 || NET8_0
 using System.Text.Json;
 #endif
 
@@ -11,7 +11,7 @@ namespace EFCoreSecondLevelCacheInterceptor;
 /// </summary>
 public class EFCoreSecondLevelCacheSettings
 {
-#if NET5_0 || NET6_0 || NET7_0 || NET8_0
+#if NET9_0 || NET5_0 || NET6_0 || NET7_0 || NET8_0
     /// <summary>
     ///     Options to control the serialization behavior
     /// </summary>
