@@ -11,53 +11,45 @@ namespace EFCoreSecondLevelCacheInterceptor;
 /// <summary>
 ///     SqlCommands Utils
 /// </summary>
-public class EFSqlCommandsProcessor : IEFSqlCommandsProcessor
+/// <remarks>
+///     SqlCommands Utils
+/// </remarks>
+public class EFSqlCommandsProcessor(IEFHashProvider hashProvider) : IEFSqlCommandsProcessor
 {
-    private static readonly string[] CrudMarkers =
-    {
-        "MERGE ", "insert ", "update ", "delete ", "create "
-    };
+    private static readonly string[] CrudMarkers = ["MERGE ", "insert ", "update ", "delete ", "create "];
 
     private static readonly Type IEntityType =
-        Type.GetType("Microsoft.EntityFrameworkCore.Metadata.IEntityType, Microsoft.EntityFrameworkCore") ??
-        throw new TypeLoadException("Couldn't load Microsoft.EntityFrameworkCore.Metadata.IEntityType");
+        Type.GetType(typeName: "Microsoft.EntityFrameworkCore.Metadata.IEntityType, Microsoft.EntityFrameworkCore") ??
+        throw new TypeLoadException(message: "Couldn't load Microsoft.EntityFrameworkCore.Metadata.IEntityType");
 
-    private static readonly PropertyInfo ClrTypePropertyInfo = IEntityType.GetInterfaces().Union(new[]
-                                                                   {
-                                                                       IEntityType
-                                                                   }).Select(i => i.GetProperty("ClrType",
-                                                                       BindingFlags.Public | BindingFlags.Instance))
-                                                                   .Distinct()
-                                                                   .FirstOrDefault(propertyInfo
-                                                                       => propertyInfo != null) ??
-                                                               throw new KeyNotFoundException(
-                                                                   "Couldn't find `ClrType` on IEntityType.");
+    private static readonly PropertyInfo ClrTypePropertyInfo =
+        IEntityType.GetInterfaces()
+            .Union([IEntityType])
+            .Select(i => i.GetProperty(name: "ClrType", BindingFlags.Public | BindingFlags.Instance))
+            .Distinct()
+            .FirstOrDefault(propertyInfo => propertyInfo != null) ??
+        throw new KeyNotFoundException(message: "Couldn't find `ClrType` on IEntityType.");
 
     private static readonly Type RelationalEntityTypeExtensionsType =
         Type.GetType(
+            typeName:
             "Microsoft.EntityFrameworkCore.RelationalEntityTypeExtensions, Microsoft.EntityFrameworkCore.Relational") ??
-        throw new TypeLoadException("Couldn't load Microsoft.EntityFrameworkCore.RelationalEntityTypeExtensions");
+        throw new TypeLoadException(
+            message: "Couldn't load Microsoft.EntityFrameworkCore.RelationalEntityTypeExtensions");
 
     private static readonly MethodInfo GetTableNameMethodInfo =
-        RelationalEntityTypeExtensionsType.GetMethod("GetTableName", BindingFlags.Static | BindingFlags.Public) ??
-        throw new KeyNotFoundException("Couldn't find `GetTableName()` on RelationalEntityTypeExtensions.");
+        RelationalEntityTypeExtensionsType.GetMethod(name: "GetTableName", BindingFlags.Static | BindingFlags.Public) ??
+        throw new KeyNotFoundException(message: "Couldn't find `GetTableName()` on RelationalEntityTypeExtensions.");
 
-    private static readonly string[] Separator =
-    {
-        "."
-    };
+    private static readonly string[] Separator = ["."];
 
     private readonly ConcurrentDictionary<string, Lazy<SortedSet<string>>> _commandTableNames =
         new(StringComparer.OrdinalIgnoreCase);
 
     private readonly ConcurrentDictionary<Type, Lazy<List<TableEntityInfo>>> _contextTableNames = new();
-    private readonly IEFHashProvider _hashProvider;
 
-    /// <summary>
-    ///     SqlCommands Utils
-    /// </summary>
-    public EFSqlCommandsProcessor(IEFHashProvider hashProvider)
-        => _hashProvider = hashProvider ?? throw new ArgumentNullException(nameof(hashProvider));
+    private readonly IEFHashProvider _hashProvider =
+        hashProvider ?? throw new ArgumentNullException(nameof(hashProvider));
 
     /// <summary>
     ///     Is `insert`, `update` or `delete`?
@@ -69,7 +61,8 @@ public class EFSqlCommandsProcessor : IEFSqlCommandsProcessor
             return false;
         }
 
-        var lines = text.Split('\n');
+        var lines = text.Split(separator: '\n');
+
         foreach (var line in lines)
         {
             foreach (var marker in CrudMarkers)
@@ -95,8 +88,9 @@ public class EFSqlCommandsProcessor : IEFSqlCommandsProcessor
         }
 
         return _contextTableNames.GetOrAdd(context.GetType(),
-            _ => new Lazy<List<TableEntityInfo>>(() => getTableNames(context),
-                LazyThreadSafetyMode.ExecutionAndPublication)).Value;
+                _ => new Lazy<List<TableEntityInfo>>(() => GetTableNames(context),
+                    LazyThreadSafetyMode.ExecutionAndPublication))
+            .Value;
     }
 
     /// <summary>
@@ -105,9 +99,11 @@ public class EFSqlCommandsProcessor : IEFSqlCommandsProcessor
     public SortedSet<string> GetSqlCommandTableNames(string commandText)
     {
         var commandTextKey = $"{_hashProvider.ComputeHash(commandText):X}";
+
         return _commandTableNames.GetOrAdd(commandTextKey,
-            _ => new Lazy<SortedSet<string>>(() => getRawSqlCommandTableNames(commandText),
-                LazyThreadSafetyMode.ExecutionAndPublication)).Value;
+                _ => new Lazy<SortedSet<string>>(() => GetRawSqlCommandTableNames(commandText),
+                    LazyThreadSafetyMode.ExecutionAndPublication))
+            .Value;
     }
 
     /// <summary>
@@ -116,53 +112,52 @@ public class EFSqlCommandsProcessor : IEFSqlCommandsProcessor
     public IList<Type> GetSqlCommandEntityTypes(string commandText, IList<TableEntityInfo> allEntityTypes)
     {
         var commandTableNames = GetSqlCommandTableNames(commandText);
+
         return allEntityTypes.Where(entityType => commandTableNames.Contains(entityType.TableName))
-            .Select(entityType => entityType.ClrType).ToList();
+            .Select(entityType => entityType.ClrType)
+            .ToList();
     }
 
-    private static List<TableEntityInfo> getTableNames(DbContext context)
+    private static List<TableEntityInfo> GetTableNames(DbContext context)
     {
         var tableNames = new List<TableEntityInfo>();
+
         foreach (var entityType in context.Model.GetEntityTypes())
         {
-            var clrType = getClrType(entityType);
+            var clrType = GetClrType(entityType);
+
             tableNames.Add(new TableEntityInfo
             {
                 ClrType = clrType,
-                TableName = getTableName(entityType) ?? clrType.ToString()
+                TableName = GetTableName(entityType) ?? clrType.ToString()
             });
         }
 
         return tableNames;
     }
 
-    private static string? getTableName(object entityType)
-        => GetTableNameMethodInfo.Invoke(null, new[]
-        {
-            entityType
-        }) as string;
+    private static string? GetTableName(object entityType)
+        => GetTableNameMethodInfo.Invoke(obj: null, [entityType]) as string;
 
-    private static Type getClrType(object entityType)
+    private static Type GetClrType(object entityType)
     {
         var value = ClrTypePropertyInfo.GetValue(entityType) ??
                     throw new InvalidOperationException($"Couldn't get the ClrType value of `{entityType}`");
+
         return (Type)value;
     }
 
-    private static SortedSet<string> getRawSqlCommandTableNames(string commandText)
+    private static SortedSet<string> GetRawSqlCommandTableNames(string commandText)
     {
-        string[] tableMarkers =
-        {
-            "FROM", "JOIN", "INTO", "UPDATE", "MERGE"
-        };
+        string[] tableMarkers = ["FROM", "JOIN", "INTO", "UPDATE", "MERGE"];
 
         var tables = new SortedSet<string>(StringComparer.OrdinalIgnoreCase);
 
-        var sqlItems = commandText.Split(new[]
-        {
-            " ", "\r\n", Environment.NewLine, "\n"
-        }, StringSplitOptions.RemoveEmptyEntries);
+        var sqlItems =
+            commandText.Split([" ", "\r\n", Environment.NewLine, "\n"], StringSplitOptions.RemoveEmptyEntries);
+
         var sqlItemsLength = sqlItems.Length;
+
         for (var i = 0; i < sqlItemsLength; i++)
         {
             foreach (var marker in tableMarkers)
@@ -173,6 +168,7 @@ public class EFSqlCommandsProcessor : IEFSqlCommandsProcessor
                 }
 
                 ++i;
+
                 if (i >= sqlItemsLength)
                 {
                     break;
@@ -181,6 +177,7 @@ public class EFSqlCommandsProcessor : IEFSqlCommandsProcessor
                 var tableName = string.Empty;
 
                 var tableNameParts = sqlItems[i].Split(Separator, StringSplitOptions.RemoveEmptyEntries);
+
                 if (tableNameParts.Length == 1)
                 {
                     tableName = tableNameParts[0].Trim();
@@ -195,9 +192,12 @@ public class EFSqlCommandsProcessor : IEFSqlCommandsProcessor
                     continue;
                 }
 
-                tableName = tableName.Replace("[", "", StringComparison.Ordinal)
-                    .Replace("]", "", StringComparison.Ordinal).Replace("'", "", StringComparison.Ordinal)
-                    .Replace("`", "", StringComparison.Ordinal).Replace("\"", "", StringComparison.Ordinal);
+                tableName = tableName.Replace(oldValue: "[", newValue: "", StringComparison.Ordinal)
+                    .Replace(oldValue: "]", newValue: "", StringComparison.Ordinal)
+                    .Replace(oldValue: "'", newValue: "", StringComparison.Ordinal)
+                    .Replace(oldValue: "`", newValue: "", StringComparison.Ordinal)
+                    .Replace(oldValue: "\"", newValue: "", StringComparison.Ordinal);
+
                 tables.Add(tableName);
             }
         }
