@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("EFCoreSecondLevelCacheInterceptor.UnitTests")]
+[assembly: InternalsVisibleTo(assemblyName: "EFCoreSecondLevelCacheInterceptor.UnitTests")]
 
 namespace EFCoreSecondLevelCacheInterceptor;
 
@@ -15,8 +15,7 @@ internal static class StringExtensions
     /// <summary>
     ///     Determines if a collection contains an item which ends with the given value
     /// </summary>
-    public static bool EndsWith(this IEnumerable<string>? collection, string? value,
-                                StringComparison stringComparison)
+    public static bool EndsWith(this IEnumerable<string>? collection, string? value, StringComparison stringComparison)
     {
         if (string.IsNullOrEmpty(value))
         {
@@ -29,8 +28,9 @@ internal static class StringExtensions
     /// <summary>
     ///     Determines if a collection contains an item which starts with the given value
     /// </summary>
-    public static bool StartsWith(this IEnumerable<string>? collection, string? value,
-                                  StringComparison stringComparison)
+    public static bool StartsWith(this IEnumerable<string>? collection,
+        string? value,
+        StringComparison stringComparison)
     {
         if (string.IsNullOrEmpty(value))
         {
@@ -43,24 +43,25 @@ internal static class StringExtensions
     /// <summary>
     ///     Determines if a collection exclusively contains every item in the given collection
     /// </summary>
-    public static bool ContainsEvery(this IEnumerable<string>? source, IEnumerable<string>? collection,
-                                     StringComparer stringComparison)
+    public static bool ContainsEvery(this IEnumerable<string>? source,
+        IEnumerable<string>? collection,
+        StringComparer stringComparison)
     {
         if (source is null || collection is null)
         {
             return false;
         }
 
-        return source.OrderBy(fElement => fElement, stringComparison).SequenceEqual(
-         collection.OrderBy(sElement => sElement, stringComparison),
-         stringComparison);
+        return source.OrderBy(fElement => fElement, stringComparison)
+            .SequenceEqual(collection.OrderBy(sElement => sElement, stringComparison), stringComparison);
     }
 
     /// <summary>
     ///     Determines if a collection contains items only in the given collection
     /// </summary>
-    public static bool ContainsOnly(this IEnumerable<string>? source, IEnumerable<string>? collection,
-                                    StringComparer stringComparison)
+    public static bool ContainsOnly(this IEnumerable<string>? source,
+        IEnumerable<string>? collection,
+        StringComparer stringComparison)
     {
         if (source is null || collection is null)
         {
@@ -69,7 +70,6 @@ internal static class StringExtensions
 
         return source.All(sElement => collection.Contains(sElement, stringComparison));
     }
-
 
 #if NET4_6_2 || NETSTANDARD2_0
         /// <summary>
@@ -83,7 +83,7 @@ internal static class StringExtensions
         /// </summary>
         public static string Replace(this string str, string oldValue, string newValue, StringComparison comparisonType)
         {
-            newValue = newValue ?? string.Empty;
+            newValue ??= string.Empty;
             if (string.IsNullOrEmpty(str) || string.IsNullOrEmpty(oldValue) || oldValue.Equals(newValue, comparisonType))
             {
                 return str;
