@@ -14,9 +14,9 @@ public class EFCacheKeyProviderTests
 {
     private readonly Mock<IEFCacheDependenciesProcessor> _cacheDependenciesProcessorMock;
     private readonly Mock<IEFCacheKeyPrefixProvider> _cacheKeyPrefixProviderMock;
+    private readonly IEFCacheKeyProvider _cacheKeyProvider;
     private readonly Mock<IEFHashProvider> _hashProviderMock;
     private readonly Mock<IEFDebugLogger> _loggerMock;
-    private readonly IEFCacheKeyProvider _cacheKeyProvider;
 
     public EFCacheKeyProviderTests()
     {
@@ -32,14 +32,9 @@ public class EFCacheKeyProviderTests
 
         cacheSettingsMock.Setup(c => c.Value).Returns(settings);
 
-        _cacheKeyProvider = new EFCacheKeyProvider(
-            _cacheDependenciesProcessorMock.Object,
-            cachePolicyParserMock.Object,
-            _loggerMock.Object,
-            keyProviderLoggerMock.Object,
-            _hashProviderMock.Object,
-            _cacheKeyPrefixProviderMock.Object,
-            cacheSettingsMock.Object);
+        _cacheKeyProvider = new EFCacheKeyProvider(_cacheDependenciesProcessorMock.Object, cachePolicyParserMock.Object,
+            _loggerMock.Object, keyProviderLoggerMock.Object, _hashProviderMock.Object,
+            _cacheKeyPrefixProviderMock.Object, cacheSettingsMock.Object);
     }
 
     [Fact]
@@ -51,22 +46,18 @@ public class EFCacheKeyProviderTests
         var loggerMock = new Mock<IEFDebugLogger>();
         var keyProviderLoggerMock = new Mock<ILogger<EFCacheKeyProvider>>();
         var cacheKeyPrefixProviderMock = new Mock<IEFCacheKeyPrefixProvider>();
-#if NET5_0 || NET6_0 || NET7_0 || NET8_0
+#if NET9_0 || NET5_0 || NET6_0 || NET7_0 || NET8_0
         var cacheSettingsMock = new Mock<IOptions<EFCoreSecondLevelCacheSettings>>();
 #endif
 
         // Act && Assert
-        Assert.Throws<ArgumentNullException>(() => new EFCacheKeyProvider(
-            cacheDependenciesProcessorMock.Object,
-            cachePolicyParserMock.Object,
-            loggerMock.Object,
-            keyProviderLoggerMock.Object,
+        Assert.Throws<ArgumentNullException>(() => new EFCacheKeyProvider(cacheDependenciesProcessorMock.Object,
+            cachePolicyParserMock.Object, loggerMock.Object, keyProviderLoggerMock.Object,
+
             // ReSharper disable once AssignNullToNotNullAttribute
-            null,
-            cacheKeyPrefixProviderMock.Object
-#if NET5_0 || NET6_0 || NET7_0 || NET8_0
-            ,
-            cacheSettingsMock.Object
+            hashProvider: null, cacheKeyPrefixProviderMock.Object
+#if NET9_0 || NET5_0 || NET6_0 || NET7_0 || NET8_0
+            , cacheSettingsMock.Object
 #endif
         ));
     }
@@ -75,7 +66,7 @@ public class EFCacheKeyProviderTests
     public void EFCacheKeyProvider_ThrowsArgumentNullException_WhenCacheSettingsIsNull()
     {
         // Arrange
-#if NET5_0 || NET6_0 || NET7_0 || NET8_0
+#if NET9_0 || NET5_0 || NET6_0 || NET7_0 || NET8_0
         var cacheDependenciesProcessorMock = new Mock<IEFCacheDependenciesProcessor>();
         var cachePolicyParserMock = new Mock<IEFCachePolicyParser>();
         var loggerMock = new Mock<IEFDebugLogger>();
@@ -84,16 +75,12 @@ public class EFCacheKeyProviderTests
         var cacheKeyPrefixProviderMock = new Mock<IEFCacheKeyPrefixProvider>();
 
         // Act && Assert
-        Assert.Throws<ArgumentNullException>(() => new EFCacheKeyProvider(
-            cacheDependenciesProcessorMock.Object,
-            cachePolicyParserMock.Object,
-            loggerMock.Object,
-            keyProviderLoggerMock.Object,
-            hashProviderMock.Object,
+        Assert.Throws<ArgumentNullException>(() => new EFCacheKeyProvider(cacheDependenciesProcessorMock.Object,
+            cachePolicyParserMock.Object, loggerMock.Object, keyProviderLoggerMock.Object, hashProviderMock.Object,
             cacheKeyPrefixProviderMock.Object,
+
             // ReSharper disable once AssignNullToNotNullAttribute
-            null
-        ));
+            cacheSettings: null));
 #endif
     }
 
@@ -107,22 +94,16 @@ public class EFCacheKeyProviderTests
         var keyProviderLoggerMock = new Mock<ILogger<EFCacheKeyProvider>>();
         var hashProviderMock = new Mock<IEFHashProvider>();
         var cacheKeyPrefixProviderMock = new Mock<IEFCacheKeyPrefixProvider>();
-#if NET5_0 || NET6_0 || NET7_0 || NET8_0
+#if NET9_0 || NET5_0 || NET6_0 || NET7_0 || NET8_0
         var cacheSettingsMock = new Mock<IOptions<EFCoreSecondLevelCacheSettings>>();
         cacheSettingsMock.Setup(c => c.Value).Returns(new EFCoreSecondLevelCacheSettings());
 #endif
 
         // Act
-        var provider = new EFCacheKeyProvider(
-            cacheDependenciesProcessorMock.Object,
-            cachePolicyParserMock.Object,
-            loggerMock.Object,
-            keyProviderLoggerMock.Object,
-            hashProviderMock.Object,
-            cacheKeyPrefixProviderMock.Object
-#if NET5_0 || NET6_0 || NET7_0 || NET8_0
-            ,
-            cacheSettingsMock.Object
+        var provider = new EFCacheKeyProvider(cacheDependenciesProcessorMock.Object, cachePolicyParserMock.Object,
+            loggerMock.Object, keyProviderLoggerMock.Object, hashProviderMock.Object, cacheKeyPrefixProviderMock.Object
+#if NET9_0 || NET5_0 || NET6_0 || NET7_0 || NET8_0
+            , cacheSettingsMock.Object
 #endif
         );
 
@@ -131,7 +112,7 @@ public class EFCacheKeyProviderTests
     }
 
     [Fact]
-    [SuppressMessage("ReSharper", "AssignNullToNotNullAttribute")]
+    [SuppressMessage(category: "ReSharper", checkId: "AssignNullToNotNullAttribute")]
     public void GetEFCacheKey_ThrowsArgumentNullException_WhenContextIsNull()
     {
         // Arrange
@@ -143,31 +124,31 @@ public class EFCacheKeyProviderTests
         void Act() => _cacheKeyProvider.GetEFCacheKey(command, context, cachePolicy);
 
         // Assert
-        Assert.Throws<ArgumentNullException>("context", Act);
+        Assert.Throws<ArgumentNullException>(paramName: "context", Act);
     }
 
     [Fact]
-    [SuppressMessage("ReSharper", "AssignNullToNotNullAttribute")]
+    [SuppressMessage(category: "ReSharper", checkId: "AssignNullToNotNullAttribute")]
     public void GetEFCacheKey_ThrowsArgumentNullException_WhenCommandIsNull()
     {
         // Arrange
         DbCommand command = null;
-        DbContext context = Mock.Of<DbContext>();
+        var context = Mock.Of<DbContext>();
         EFCachePolicy cachePolicy = null;
 
         // Act
         void Act() => _cacheKeyProvider.GetEFCacheKey(command, context, cachePolicy);
 
         // Assert
-        Assert.Throws<ArgumentNullException>("command", Act);
+        Assert.Throws<ArgumentNullException>(paramName: "command", Act);
     }
 
     [Fact]
     public void GetEFCacheKey_ThrowsArgumentNullException_WhenCachePolicyIsNull()
     {
         // Arrange
-        DbCommand command = Mock.Of<DbCommand>();
-        DbContext context = Mock.Of<DbContext>();
+        var command = Mock.Of<DbCommand>();
+        var context = Mock.Of<DbContext>();
         EFCachePolicy cachePolicy = null;
 
         // Act
@@ -175,7 +156,7 @@ public class EFCacheKeyProviderTests
         void Act() => _cacheKeyProvider.GetEFCacheKey(command, context, cachePolicy);
 
         // Assert
-        Assert.Throws<ArgumentNullException>("cachePolicy", Act);
+        Assert.Throws<ArgumentNullException>(paramName: "cachePolicy", Act);
     }
 
     [Fact]
@@ -183,6 +164,7 @@ public class EFCacheKeyProviderTests
     {
         // Arrange
         var context = Mock.Of<DbContext>();
+
         var expected = new EFCacheKey(new SortedSet<string>())
         {
             KeyHash = "75BCD15",
@@ -190,37 +172,33 @@ public class EFCacheKeyProviderTests
         };
 
         var commandMock = new Mock<DbCommand>();
-        var cachePolicy = new EFCachePolicy().SaltKey("CacheSaltKey");
+        var cachePolicy = new EFCachePolicy().SaltKey(saltKey: "CacheSaltKey");
         var dbParameterCollectionMock = new Mock<DbParameterCollection>();
         var dbParameterMock = new Mock<DbParameter>();
+
         var parameters = new List<DbParameter>
         {
-            dbParameterMock.Object,
+            dbParameterMock.Object
         };
 
-        _hashProviderMock
-            .Setup(x => x.ComputeHash(@"ConnectionString
+        _hashProviderMock.Setup(x => x.ComputeHash(@"ConnectionString
 =Name=""Value"",Size=2147483647,Precision=255,Scale=255,Direction=Input,SaltKey
 =CacheSaltKey"))
-            .Returns(123456789);
+            .Returns(value: 123456789);
 
-        _loggerMock.Setup(x => x.IsLoggerEnabled).Returns(true);
+        _loggerMock.Setup(x => x.IsLoggerEnabled).Returns(value: true);
 
-        commandMock
-            .Protected()
-            .SetupGet<DbParameterCollection>("DbParameterCollection")
+        commandMock.Protected()
+            .SetupGet<DbParameterCollection>(propertyName: "DbParameterCollection")
             .Returns(dbParameterCollectionMock.Object);
 
-        dbParameterCollectionMock
-            .Setup(x => x.GetEnumerator())
-            .Returns(parameters.GetEnumerator());
+        dbParameterCollectionMock.Setup(x => x.GetEnumerator()).Returns(parameters.GetEnumerator());
 
-        _cacheDependenciesProcessorMock
-            .Setup(x => x.GetCacheDependencies(commandMock.Object, context, cachePolicy))
+        _cacheDependenciesProcessorMock.Setup(x => x.GetCacheDependencies(commandMock.Object, context, cachePolicy))
             .Returns(expected.CacheDependencies as SortedSet<string>);
 
-        dbParameterMock.Setup(x => x.ParameterName).Returns("Name");
-        dbParameterMock.Setup(x => x.Value).Returns("Value");
+        dbParameterMock.Setup(x => x.ParameterName).Returns(value: "Name");
+        dbParameterMock.Setup(x => x.Value).Returns(value: "Value");
         dbParameterMock.Setup(x => x.Size).Returns(int.MaxValue);
         dbParameterMock.Setup(x => x.Precision).Returns(byte.MaxValue);
         dbParameterMock.Setup(x => x.Scale).Returns(byte.MaxValue);
@@ -238,6 +216,7 @@ public class EFCacheKeyProviderTests
     {
         // Arrange
         var context = Mock.Of<DbContext>();
+
         var expected = new EFCacheKey(new SortedSet<string>())
         {
             KeyHash = "CacheKeyPrefix75BCD15",
@@ -245,41 +224,35 @@ public class EFCacheKeyProviderTests
         };
 
         var commandMock = new Mock<DbCommand>();
-        var cachePolicy = new EFCachePolicy().SaltKey("CacheSaltKey");
+        var cachePolicy = new EFCachePolicy().SaltKey(saltKey: "CacheSaltKey");
         var dbParameterCollectionMock = new Mock<DbParameterCollection>();
         var dbParameterMock = new Mock<DbParameter>();
+
         var parameters = new List<DbParameter>
         {
-            dbParameterMock.Object,
+            dbParameterMock.Object
         };
 
-        _cacheKeyPrefixProviderMock
-            .Setup(x => x.GetCacheKeyPrefix())
-            .Returns("CacheKeyPrefix");
+        _cacheKeyPrefixProviderMock.Setup(x => x.GetCacheKeyPrefix()).Returns(value: "CacheKeyPrefix");
 
-        _hashProviderMock
-            .Setup(x => x.ComputeHash(@"ConnectionString
+        _hashProviderMock.Setup(x => x.ComputeHash(@"ConnectionString
 =Name=""Value"",Size=2147483647,Precision=255,Scale=255,Direction=Input,SaltKey
 =CacheSaltKey"))
-            .Returns(123456789);
+            .Returns(value: 123456789);
 
-        _loggerMock.Setup(x => x.IsLoggerEnabled).Returns(true);
+        _loggerMock.Setup(x => x.IsLoggerEnabled).Returns(value: true);
 
-        commandMock
-            .Protected()
-            .SetupGet<DbParameterCollection>("DbParameterCollection")
+        commandMock.Protected()
+            .SetupGet<DbParameterCollection>(propertyName: "DbParameterCollection")
             .Returns(dbParameterCollectionMock.Object);
 
-        dbParameterCollectionMock
-            .Setup(x => x.GetEnumerator())
-            .Returns(parameters.GetEnumerator());
+        dbParameterCollectionMock.Setup(x => x.GetEnumerator()).Returns(parameters.GetEnumerator());
 
-        _cacheDependenciesProcessorMock
-            .Setup(x => x.GetCacheDependencies(commandMock.Object, context, cachePolicy))
+        _cacheDependenciesProcessorMock.Setup(x => x.GetCacheDependencies(commandMock.Object, context, cachePolicy))
             .Returns(expected.CacheDependencies as SortedSet<string>);
 
-        dbParameterMock.Setup(x => x.ParameterName).Returns("Name");
-        dbParameterMock.Setup(x => x.Value).Returns("Value");
+        dbParameterMock.Setup(x => x.ParameterName).Returns(value: "Name");
+        dbParameterMock.Setup(x => x.Value).Returns(value: "Value");
         dbParameterMock.Setup(x => x.Size).Returns(int.MaxValue);
         dbParameterMock.Setup(x => x.Precision).Returns(byte.MaxValue);
         dbParameterMock.Setup(x => x.Scale).Returns(byte.MaxValue);
